+++ src/matrix/ft/CSENSE.hpp
@@ -133,7 +133,7 @@ public:
 			
 			// FT individual channels
 #pragma omp for
-			for (int i = 0; i < nc; i++)
+			for (int i = 0; i < (int)nc; i++)
 				if (ndim == 2)
 					Slice  (tmp, i, ft ->* Slice  (tmp, i));
 				else
@@ -141,16 +141,16 @@ public:
 
 			// Antialiasing
 #pragma omp for
-			for (int x = 0; x < dims[0]; x++)
-				for (int y = 0; y < dims[1]; y++)
-					for (int z = 0; z < dims[2]; z++) {
-						for (int c = 0; c < nc; c++) {
+			for (int x = 0; x < (int)dims[0]; x++)
+				for (size_t y = 0; y < dims[1]; y++)
+					for (size_t z = 0; z < dims[2]; z++) {
+						for (size_t c = 0; c < nc; c++) {
 
 							ra[c] = (ndim == 3) ? tmp (x, y, z, c) : tmp (x, y, c);
 							
-							for (int zi = 0, i = 0; zi < af[2]; zi++)
-								for (int yi = 0; yi < af[1]; yi++)
-									for (int xi = 0; xi < af[0]; xi++, i++) 
+							for (size_t zi = 0, i = 0; zi < af[2]; zi++)
+								for (size_t yi = 0; yi < af[1]; yi++)
+									for (size_t xi = 0; xi < af[0]; xi++, i++) 
                                         s (c, i) = (ndim == 3) ?
 											sens (x + xi * dims[0], y + yi * dims[1], z + zi * dims[2], c):
 											sens (x + xi * dims[0], y + yi * dims[1],                   c);
@@ -176,9 +176,9 @@ public:
 						si = gemm (si,  s, 'N', 'C');
 							rp = gemm (si, ra, 'N', 'N');
                         
-                        for (int zi = 0, i = 0; zi < af[2]; zi++)
-                            for (int yi = 0; yi < af[1]; yi++)
-                                for (int xi = 0; xi < af[0]; xi++, i++) {
+                        for (size_t zi = 0, i = 0; zi < af[2]; zi++)
+                            for (size_t yi = 0; yi < af[1]; yi++)
+                                for (size_t xi = 0; xi < af[0]; xi++, i++) {
 #if defined (_MSC_VER) && (_MSC_VER < 1600)
 									if (ndim == 3)
 										res (x + xi * dims[0], y + yi * dims[1], z + zi * dims[2], 0) = rp [i];
